Move updateRun to the seperate file
related #39
