+++ src/database/db-add-run.js
@@ -1,22 +1,6 @@
 const GoodGymDB = require('./db-connection');
 const findOneRun = require('./db-find-one');
-
-const updateRun = (inputRun) => {
-  const updateOption = {
-    '$set': {
-      'run.task': inputRun.task,
-      'run.location': inputRun.location,
-      'run.purpose': inputRun.purpose,
-      'run.contact': inputRun.contact,
-      'run.risk': inputRun.risk
-    }
-  };
-
-  GoodGymDB.findOneAndUpdate({'run.runId': inputRun.runId}, updateOption, (err, run) => {
-    if (err) throw err;
-    console.log(run, 'Updated run!');
-  });
-};
+const updateRun = require('./db-update-run');
 
 const addRunToDB = (inputRun, cb) => {
   let newRun = GoodGymDB({
